+++ pyrax/cloudloadbalancers.py
@@ -488,8 +488,8 @@ class CloudLoadBalancerManager(BaseManager):
         if not all(required):
             raise exc.MissingLoadBalancerParameters("Load Balancer creation "
                 "requires at least one virtual IP, a protocol, and a port.")
-        nodes = utils.coerce_string_to_list(nodes)
-        virtual_ips = utils.coerce_string_to_list(virtual_ips)
+        nodes = utils.coerce_to_list(nodes)
+        virtual_ips = utils.coerce_to_list(virtual_ips)
         bad_conditions = [node.condition for node in nodes
                 if node.condition.upper() not in ("ENABLED", "DISABLED")]
         if bad_conditions: